修复：校正兼容入口类型约束
为凭据持久化路径保留显式生命周期，以闭包读取青龙环境变量，并将 DaCapo 签到尝试次数统一为 u32。

+++ src/config/dacapo.rs
@@ -79,7 +79,7 @@ fn parse_dacapo(source: &str, account_name: &str) -> Result<LoadedConfig, Dacapo
         "国服游戏.重试次数",
         3,
     )?;
-    let retries = u8::try_from(retries).map_err(|_| DacapoError::Field {
+    let retries = u32::try_from(retries).map_err(|_| DacapoError::Field {
         field: "国服游戏.重试次数",
         message: "必须在 1 到 10 之间",
     })?;
